First GET/POST methods implemented

+++ src/main/java/com/fujitsu/drugsapp/controllers/DrugSetController.java
@@ -2,6 +2,7 @@ package com.fujitsu.drugsapp.controllers;
 
 import com.fujitsu.drugsapp.entities.Drug;
 import com.fujitsu.drugsapp.entities.DrugSet;
+import com.fujitsu.drugsapp.entities.DrugUpdate;
 import com.fujitsu.drugsapp.services.DrugSetService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,16 +40,16 @@ public class DrugSetController {
         return new ResponseEntity<>(drugSetService.findById(UUID.fromString(id)), HttpStatus.OK);
     }
 
+    @GetMapping("/{id}/updates")
+    public ResponseEntity<List<DrugUpdate>> getDrugSetUpdates(@PathVariable("id") String id) {
+        return new ResponseEntity<>(drugSetService.getDrugSetUpdates(UUID.fromString(id)), HttpStatus.OK);
+    }
+
     @GetMapping("/{id}/drugs")
-    public ResponseEntity<List<Drug>> getDrugsInDrugSet(@PathVariable("id") String id, @RequestParam(required = false) String searchText, @RequestParam(name = "date", required = false) Instant date) {
+    public ResponseEntity<List<Drug>> getDrugsInDrugSet(@PathVariable("id") String id, @RequestParam(required = false) String searchText,
+                                                        @RequestParam(name = "date", required = false) Instant date) {
 
-        if(date!=null) {
-            LocalDateTime localDateTime = LocalDateTime.ofInstant(date, ZoneOffset.UTC);
-            return new ResponseEntity<List<Drug>>(drugSetService.findDrugsById(UUID.fromString(id), searchText, localDateTime), HttpStatus.OK);
-        }else{
-            return new ResponseEntity<List<Drug>>(drugSetService.findDrugsById(UUID.fromString(id), searchText, LocalDateTime.ofInstant(date, ZoneOffset.UTC)), HttpStatus.OK);
-        }
-
+        return new ResponseEntity<List<Drug>>(drugSetService.findDrugsById(UUID.fromString(id), searchText, date), HttpStatus.OK);
     }
 
     @PostMapping("/new")
@@ -57,17 +58,20 @@ public class DrugSetController {
         return new ResponseEntity<>(drugSetService.saveDrugSet(drugSet), HttpStatus.CREATED);
     }
 
-    @PutMapping("/{id}/update")
-    public ResponseEntity<DrugSet> updateDrugSet(@PathVariable("id") String id, @RequestBody DrugSet drugSet) {
+    @PostMapping("/{id}/update")
+    public ResponseEntity<DrugUpdate> updateDrugSet(@PathVariable("id") String id) {
         if (drugSetService.existById(UUID.fromString(id))) {
-            return new ResponseEntity<>(drugSetService.saveDrugSet(drugSet), HttpStatus.ACCEPTED);
+            return new ResponseEntity<>(drugSetService.updateDrugSet(UUID.fromString(id)), HttpStatus.ACCEPTED);
         }
         throw new IllegalArgumentException("DrugSet with id " + id + "not found");
     }
     @DeleteMapping("/{id}/delete")
     public ResponseEntity<HttpStatus> deleteDrugSet(@PathVariable("id") String id) {
-        drugSetService.deleteDrugSet(UUID.fromString(id));
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if (drugSetService.existById(UUID.fromString(id))) {
+            drugSetService.deleteDrugSet(UUID.fromString(id));
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        throw new IllegalArgumentException("DrugSet with id " + id + "not found");
     }
 
 }
